fix(perf_trigger): support new format & reduce output size

+++ perf_trigger/modules/gcpt.py
@@ -17,7 +17,7 @@ class GCPT:
         benchmark: str,
         checkpoint: str,
         weight: float,
-        eta: float = 0.0,
+        eta: int = 0,
     ):
         self.__gcpt_path = gcpt_path
         self.__benchmark = benchmark
